Add check for content-length before allocating buffers
Currently, the `to_bytes` function prevents a remote to send too much
data. But the functions allocate a buffer if the request contains more
than a chunk. A size hint is used to allocate with capacity, but the
size hint takes the remote controlled `Content-Length` header into
account.

We use `to_bytes` in to places, which now are guardedby a check for
the upper size hint. This avoids readingany chunk if the
`Content-Length` header suggests therequest is larger than the library
user allows.